created product feature context class
