+++ tests/Integration/Behaviour/behat.yml
@@ -176,4 +176,5 @@ default:
                 features: Features/Scenario/Product
             contexts:
                 - Tests\Integration\Behaviour\Features\Context\CommonFeatureContext
+                - Tests\Integration\Behaviour\Features\Context\ProductFeatureContext
                 - Tests\Integration\Behaviour\Features\Context\Domain\ProductFeatureContext
